+++ Libplanet.Extensions.Cocona/Commands/KeyCommand.cs
@@ -4,10 +4,16 @@ using System.IO;
 using System.Linq;
 using global::Cocona;
 using Libplanet.Crypto;
+using Libplanet.Extensions.Cocona.Commands.Key;
 using Libplanet.KeyStore;
 
 namespace Libplanet.Extensions.Cocona.Commands
 {
+    [HasSubCommands(
+        typeof(DerivationCommand),
+        "derive",
+        Description = "Derive the address or public key from private key."
+    )]
     public class KeyCommand
     {
         public KeyCommand()